# Optimize materializeAllLabels to Avoid Unnecessary Column Materialization