--- a/.vscode/settings.json
+++ b/.vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "go.testTimeout": "300s",
+    "go.testFlags": ["-timeout=300s"]
+}