--- a/Makefile
+++ b/Makefile
@@ -11,4 +11,4 @@ all-tests-with-coverage:
 	$(GOTEST) -timeout 600s -v -count=1 -cover -coverprofile=coverage.txt ./...
 
 bench-select:
-	$(GOTEST) -bench=BenchmarkSelect -run='^$$' ./search/... -benchtime=1s -count=1
+	$(GOTEST) -bench=BenchmarkSelect -run='^$$' ./queryable/... -benchtime=1s -count=1