# Problem Statement
The previous implementation had a performance issue specially with series with sparse labels:

The implementation before was the follow: 

* Created a global colsMap containing all column indexes present in any row across all row ranges
* Used this global map to materialize columns for every row range
* Result: Columns were being materialized even when they weren't present in specific row ranges, causing unnecessary I/O and processing overhead

This PR modify this logic to map column indexes per row range before materialization and doing so, for a given row range, we only materialize columns that is present in at least 1 series on that range.