@alanprot 2025-08-06
Bench results:

We do not see a big difference here as all series on the benchmark test has the same labels keys. But not having a worst perf (so no regression) is a good news.

```
 us-east-1:beta-cluster ➜  ~ ~/go/bin/benchstat /tmp/old /tmp/new
goos: darwin
goarch: arm64
pkg: github.com/prometheus-community/parquet-common/queryable
cpu: Apple M1 Pro
                                        │   /tmp/old   │               /tmp/new               │
                                        │    sec/op    │    sec/op     vs base                │
Select/SingleMetricAllSeries-10           466.1m ± ∞ ¹   437.6m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10       22.31m ± ∞ ¹   21.13m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10           19.93m ± ∞ ¹   19.47m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10        45.76m ± ∞ ¹   42.55m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10               18.33m ± ∞ ¹   19.41m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsRange-10             1.961 ± ∞ ¹    1.861 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10           570.4m ± ∞ ¹   512.1m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10       386.5m ± ∞ ¹   364.9m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10     1.166 ± ∞ ¹    1.159 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10      59.93m ± ∞ ¹   63.34m ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10   179.8m ± ∞ ¹   188.6m ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                   151.2m         147.4m        -2.49%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                        │      /tmp/old      │                  /tmp/new                  │
                                        │ bytes_get_range/op │ bytes_get_range/op  vs base                │
Select/SingleMetricAllSeries-10                 5.678M ± ∞ ¹         5.678M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10             327.4k ± ∞ ¹         327.4k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10                 280.5k ± ∞ ¹         280.5k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10              1.976M ± ∞ ¹         1.976M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10                     270.6k ± ∞ ¹         270.6k ± ∞ ¹       ~ (p=1.000 n=1) ³
Select/MultipleMetricsRange-10                  30.84M ± ∞ ¹         30.84M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10                 14.62M ± ∞ ¹         14.62M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10             6.330M ± ∞ ¹         6.330M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10          27.39M ± ∞ ¹         27.39M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10            2.220M ± ∞ ¹         2.220M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10         14.74M ± ∞ ¹         14.74M ± ∞ ¹       ~ (p=1.000 n=1) ³
geomean                                         3.405M               3.405M        -0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal

                                        │  /tmp/old   │              /tmp/new               │
                                        │   get/op    │   get/op     vs base                │
Select/SingleMetricAllSeries-10           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10       0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10        0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10               0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsRange-10            0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10           0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10       0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10    0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10      0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10   0.000 ± ∞ ¹   0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                             ³                +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                                        │   /tmp/old   │               /tmp/new               │
                                        │ get_range/op │ get_range/op  vs base                │
Select/SingleMetricAllSeries-10           5.400k ± ∞ ¹   5.400k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10        654.0 ± ∞ ¹    654.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10            624.0 ± ∞ ¹    624.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10        3.000k ± ∞ ¹   3.000k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10                600.0 ± ∞ ¹    600.0 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsRange-10            58.01k ± ∞ ¹   58.01k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10           43.61k ± ∞ ¹   43.61k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10       12.11k ± ∞ ¹   12.11k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10    73.33k ± ∞ ¹   73.33k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10      8.459k ± ∞ ¹   8.459k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10   62.10k ± ∞ ¹   62.10k ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                   7.570k         7.570k        +0.00%
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal

                                        │   /tmp/old   │               /tmp/new               │
                                        │  series/op   │  series/op    vs base                │
Select/SingleMetricAllSeries-10           300.0k ± ∞ ¹   300.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10       3.000k ± ∞ ¹   3.000k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10            1.000 ± ∞ ¹    1.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10        5.000k ± ∞ ¹   5.000k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10                0.000 ± ∞ ¹    0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsRange-10            1.200M ± ∞ ¹   1.200M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10           300.0k ± ∞ ¹   300.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10       234.0k ± ∞ ¹   234.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10    702.0k ± ∞ ¹   702.0k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10      6.000k ± ∞ ¹   6.000k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10    0.000 ± ∞ ¹    0.000 ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                              ³                 +0.00%               ³
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean

                                        │   /tmp/old    │               /tmp/new                │
                                        │     B/op      │     B/op       vs base                │
Select/SingleMetricAllSeries-10           824.3Mi ± ∞ ¹   785.0Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10       71.01Mi ± ∞ ¹   70.61Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10           65.84Mi ± ∞ ¹   65.84Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10        131.1Mi ± ∞ ¹   131.9Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10               62.94Mi ± ∞ ¹   62.94Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsRange-10            3.297Gi ± ∞ ¹   3.143Gi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10           1.080Gi ± ∞ ¹   1.042Gi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10       752.8Mi ± ∞ ¹   722.2Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10    2.280Gi ± ∞ ¹   2.190Gi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10      192.6Mi ± ∞ ¹   191.8Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10   539.9Mi ± ∞ ¹   539.9Mi ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                   376.7Mi         369.3Mi        -1.96%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

                                        │   /tmp/old    │               /tmp/new               │
                                        │   allocs/op   │  allocs/op    vs base                │
Select/SingleMetricAllSeries-10            9.792M ± ∞ ¹   7.996M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricReducedSeries-10        248.6k ± ∞ ¹   230.6k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricOneSeries-10            225.2k ± ∞ ¹   225.2k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/SingleMetricSparseSeries-10         593.8k ± ∞ ¹   567.5k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NonExistentSeries-10                150.9k ± ∞ ¹   150.9k ± ∞ ¹       ~ (p=1.000 n=1) ³
Select/MultipleMetricsRange-10             39.65M ± ∞ ¹   32.47M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/MultipleMetricsSparse-10           10.805M ± ∞ ¹   9.010M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexSingleMetric-10        8.303M ± ∞ ¹   6.903M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/NegativeRegexMultipleMetrics-10     24.47M ± ∞ ¹   20.27M ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexSingleMetric-10       876.5k ± ∞ ¹   840.6k ± ∞ ¹       ~ (p=1.000 n=1) ²
Select/ExpensiveRegexMultipleMetrics-10    1.796M ± ∞ ¹   1.796M ± ∞ ¹       ~ (p=1.000 n=1) ²
geomean                                    2.228M         2.013M        -9.67%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
³ all samples are equal
```